app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
httptools==0.6.1
idna==3.7
ObjDict==0.4.4
pydantic==2.7.0
pydantic_core==2.18.1
pymongo==4.6.3
PyMySQL==1.1.0
python-dotenv==0.20.0
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.2
uuid==1.30
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
